code/requirements.txt
alembic==0.8.10
apache-airflow==1.8.2
appnope==0.1.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
croniter==0.3.19
decorator==4.1.2
dill==0.2.7.1
docutils==0.14
Flask==0.11.1
Flask-Admin==1.4.1
Flask-Cache==0.13.1
Flask-Login==0.2.11
flask-swagger==0.2.13
Flask-WTF==0.14
funcsigs==1.0.0
future==0.15.2
gitdb2==2.0.3
GitPython==2.1.7
gunicorn==19.3.0
idna==2.6
ipython==6.2.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.8.1
jsonlines==1.2.0
lockfile==0.12.2
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
nltk==3.2.5
numpy==1.13.1
ordereddict==1.1
pandas==0.20.3
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psutil==4.4.2
ptyprocess==0.5.2
Pygments==2.2.0
python-daemon==2.1.2
python-dateutil==2.6.1
python-editor==1.0.3
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2017.2
PyYAML==3.12
requests==2.18.4
setproctitle==1.1.10
simplegeneric==0.8.1
six==1.11.0
smmap2==2.0.3
SQLAlchemy==1.1.14
tabulate==0.7.7
textblob==0.13.0
thrift==0.9.3
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
WTForms==2.1
zope.deprecation==4.3.0
